% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ria.R
\name{extract_ria}
\alias{extract_ria}
\title{Radiomic Feature Extraction}
\usage{
extract_ria(phasefile, leslabels, disc = T)
}
\arguments{
\item{phasefile}{Location of the T2*-phase image}

\item{leslabels}{Lesion label map, generated by our lesion identification method}

\item{disc}{Calculate discretized versions of first order radiomic features?}
}
\value{
A dataframe containing radiomic features for each identified lesion
}
\description{
This function extracts radiomic features from a given T2*-phase image and a lesion label map
}
