% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findprls.R
\name{findprls}
\alias{findprls}
\title{MS Lesion Identification and PRL Classification with Pre-trained Model}
\usage{
findprls(probmap, lesmask = probmap > thresh, phasefile,
  predmodel = pretrainedmodel, disc = T)
}
\arguments{
\item{probmap}{Lesion probability map. We recommend using lesion segmentation algorithm MIMoSA.}

\item{lesmask}{Lesion segmentation mask. Given a probability threshold, automatically binarizes lesion probability map into a segmentation mask.}

\item{phasefile}{Location of the T2*-phase image}

\item{predmodel}{Model used to generate classifications. This automatically uses a pre-trained model ported with the package.}

\item{disc}{Calculate discretized versions of first order radiomic features?}
}
\value{
A list with 3 objects: a NIfTI file of a lesion label map, a dataframe containing all radiomic features, and a vector containing lesion-wise probability of being a PRL
}
\description{
This function takes in a lesion probability map, a lesion segmentation mask, and a T2*-phase image for a single subject and identifies + classifies lesions as PRL or not.
}
