% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lesion_labels.R
\name{get_lesion_labels}
\alias{get_lesion_labels}
\title{Lesion Labeling}
\usage{
get_lesion_labels(lesion_mask, centers)
}
\arguments{
\item{lesion_mask}{Lesion mask.}

\item{centers}{Lesion center map. Provided by lesiontools::lesion_centers().}
}
\value{
A NIfTI with each lesion assigned to its closest lesion center
}
\description{
This function is a helper function for lesion_identification(). It takes in a lesion segmentation mask and a NIfTI image with identified lesion centers.
}
