% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_lesion.R
\name{label_lesion}
\alias{label_lesion}
\title{This function labels lesions in an image based on the provided probability map and a binary mask.}
\usage{
label_lesion(prob_map, bin_map, mincluster = 100)
}
\arguments{
\item{prob_map}{A probability map representing the likelihood of lesions.}

\item{bin_map}{A binary mask identifying lesions above a threshold.}

\item{mincluster}{The minimum cluster size for lesions (default is 100).}
}
\value{
A labeled image with each lesion identified. Confluent lesions are split
}
\description{
This function takes a probability map and a binary mask as input and labels lesions in the mask
based on the probability map and minimum cluster size criteria. It returns a labeled image with
lesions identified and labeled.
}
\examples{
\dontrun{
prob_map <- check_ants("prob_map.nii.gz")
mimosa_mask <- check_ants("mimosa_mask.nii.gz")
labeled_image <- label_lesion(prob_map, mimosa_mask, mincluster = 100)
}

}
\seealso{
\code{\link{lesion_centers}} for identifying lesion centers.
\code{\link{split_confluent}} for splitting confluent lesions.
}
