% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesion_centers.R
\name{lesion_centers}
\alias{lesion_centers}
\title{Distinct Lesion Centers}
\usage{
lesion_centers(prob_map, bin_map, minCenterSize = 10, radius = 1)
}
\arguments{
\item{prob_map}{an image of class \code{antsImage} containing the probability that each voxel is a lesion voxel}

\item{bin_map}{a binarized image of class \code{antsImage} in which voxels are classified as either lesion voxels or not lesion voxels.}

\item{minCenterSize}{an integer value representing the minimum number of connected voxels that can be considered a lesion center}

\item{radius}{an integer specifying radius of the neighborhood (in voxels) for which the hessian should be calculated.}
}
\value{
A list containing lesioncenters (antsImage object with labeled lesion centers) and lesioncount (an integer value representing the number of distinct lesions)
}
\description{
This function finds the centers of distinct lesions based on a lesion probability map. Each lesion center is labeled via a unique integer. The method is described in Dworkin et al., (2018).
}
\examples{
\dontrun{
library(neurobase)
lesion.probs <- readnii("path/to/probabilitymap")
centers <- lesioncenters(
  probmap = lesion.probs, binmap = lesion.probs > 0.30
)
}
}
\references{
J.D. Dworkin, K.A. Linn, I. Oguz, G.M. Fleishman, R. Bakshi, G. Nair, P.A. Calabresi, R.G. Henry, J. Oh, N. Papinutto, D. Pelletier, W. Rooney, W. Stern, N.L. Sicotte, D.S. Reich, R.T. Shinohara. An automated statistical technique for counting distinct multiple sclerosis lesions. American Journal of Neuroradiology, 2018; 39, 626-633.
}
