% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prlmodel_orig.R
\docType{data}
\name{prlmodel_orig}
\alias{prlmodel_orig}
\title{Model for classifying pre-identified lesions as PRL or not PRL}
\usage{
prlmodel_orig
}
\description{
This version is built with only undiscretized first order features.
}
\keyword{model}
\keyword{pretrained}
