% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prlmodel_orig_disc8_disc64.R
\docType{data}
\name{prlmodel_orig_disc8_disc64}
\alias{prlmodel_orig_disc8_disc64}
\title{Model for classifying pre-identified lesions as PRL or not PRL}
\usage{
prlmodel_orig_disc8_disc64
}
\description{
This version is built with undiscretized and discretized first order features.
}
\keyword{model}
\keyword{pretrained}
