% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_confluent.R
\name{split_confluent}
\alias{split_confluent}
\title{Split Confluent Lesions}
\usage{
split_confluent(i, labeled_image, lesion_center_image)
}
\arguments{
\item{i}{The label of the confluent lesions to be split.}

\item{labeled_image}{The labeled image containing lesions.}

\item{lesion_center_image}{The image containing lesion center information.}
}
\value{
A labeled image with split lesions.
}
\description{
This function takes a labeled image, lesion center image, and the label "i" of the confluent lesion
to be split. It uses the lesion center information to split the confluent lesions and returns a labeled
image with the split lesions labeled distinctly.
}
\details{
This function splits confluent lesions labeled as "i" in the labeled image
using the lesion center information and returns a labeled image with split lesions.
}
\examples{
\dontrun{
labeled_image <- check_ants("labeled_image.nii.gz")
lesion_center_image <- check_ants("lesion_center_image.nii.gz")
split_lesion_image <- split_confluent(i = 1, labeled_image, lesion_center_image)
}

}
\seealso{
\code{\link{get_lesion_labels}} for obtaining lesion labels based on lesion centers.
}
