% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_prls.R
\name{findprls}
\alias{findprls}
\title{MS Lesion Identification and PRL Classification with Pre-trained Model}
\usage{
findprls(
  phase_path,
  prob_map,
  bin_map,
  register_maps = F,
  prob_map_image = NULL,
  epi_image = NULL,
  verbose = F
)
}
\arguments{
\item{phase_path}{Filepath of T2*-phase image with appropriate preprocessing already done. Cannot be antsImage or nifti object.}

\item{prob_map}{Lesion probability map. We recommend using lesion segmentation algorithm MIMoSA. Can be path to .nii.gz, antsImage or nifti object}

\item{bin_map}{Lesion segmentation mask obtained by thresholding prob_map. Can be path to .nii.gz, antsImage or nifti object}

\item{register_maps}{Logical indicating whether prob_map and bin_map need to be registered to the phase image. (Default = F)}

\item{prob_map_image}{If register_maps == T, an image in the prob_map/bin_map space must be provided}

\item{epi_image}{If register_maps == T, an EPI image in the same space as the phase image must be provided. Registration to phase does not work well due to noise.}

\item{verbose}{Logical indicating whether print statements should be made (Default = F)}
}
\value{
A list with 2 objects: a NIfTI file of a lesions where each lesion is labeled with its probability of being a PRL and a dataframe containing all radiomic features.
}
\description{
This function takes in a lesion probability map, a lesion segmentation mask, and a T2*-phase image for a single subject and identifies + classifies lesions as PRL or not.
}
